==> notebooks/1.read_resume.ipynb <==
import os


from google.colab import drive
drive.mount('/content/drive')

#keeping all file path in dictionary
from collections import defaultdict
pdf_path_dic=defaultdict(list)

#path of data
path='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/data'
for root, dirs, files in os.walk('data'):
    for file in files:
        file_path = os.path.join(root, file)
        if file[-4:]=='.pdf':
            path=root.split('\\')
            category,file_id=path[-1],file[:-4]
            path.append(file)
            pdf_path_dic[category].append((file_id,'/'.join(path)))




pdf_path_dic['ACCOUNTANT'][:5]

pdf_path_dic['ARTS'][:5]

* reading pdf's usign PyPDF2

# import PyPDF2
import pandas as pd

#creating a df to store all data related of pdf
col_names=['id','number_of_pages','text','category']
main_df=pd.DataFrame(columns=col_names)
main_df

count=0
for category,files in pdf_path_dic.items():
    for file_id,path in files:
        pdf_file=open(path,'rb')
        pdf_reader=PyPDF2.PdfReader(pdf_file)

        if pdf_reader.is_encrypted:
            print(f'{id }this file is encrypted')
            continue

        no_of_pages=len(pdf_reader.pages)
        text=''

        for  i in range(no_of_pages):
            page=pdf_reader.pages[i]
            text+="\n"+page.extract_text()

        #creating a new df then concat with old, update the old df
        new_row_df=pd.DataFrame({'id':int(file_id),'number_of_pages':no_of_pages,'text':text,'category':category},index=[0])
        main_df=pd.concat([main_df,new_row_df],axis=0,ignore_index=True)


        count+=1




count

main_df.head()

def get_jobrole(x):
    x=x[:100]
    import re
    x=re.sub(' +',' ',x)
    l=x.split('\n')
    for i in l:
        if len(i)>3:
            return i

get_jobrole(' \nSENIOR ACCOUNTANT\nProfessional Summary\n')

main_df['job_role']=main_df['text'].apply(get_jobrole)

main_df.info()

main_df.head()

main_df['job_role'].value_counts()

save the csv reading every time pdf file is time taking

main_df.to_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/main_df.csv',index=False)

def get_index_with_id(file_id):
    for i in range(0,len(main_df['id'])):
        if main_df.loc[i]['id']==file_id:
            return i


main_df.loc[get_index_with_id(12092347)]['text']

main_df.loc[get_index_with_id(11637468)]['text']

preprocess the text

*   clean the text
*   generate tokens



import spacy
nlp = spacy.load("en_core_web_lg")


import re


def preprocess(text):
    # convert all characters in the string to lower case
    text = re.sub(r'\n', ' ', text)

    #remove digits
    text = re.sub(r'\d', '', text)

    # Remove punctuation (you can customize this pattern for your needs)
    text = re.sub(r'[^\w\s]', '', text)

    # Remove extra spaces and replace with a single space
    text = re.sub(r'\s+', ' ', text).strip()

    # tokenize word
    doc=nlp(text)
    toks=[i.text.lower() for i in doc if not i.is_stop]


    return " " .join(toks)

x=preprocess(main_df.loc[get_index_with_id(11637468)]['text'])

x

main_df['pre_text']=main_df['text'].apply(preprocess)

#better store as string instead of list
main_df['tokens']=main_df['pre_text'].apply(lambda x:','.join(x.split()))

save the csv file because reading every time from pdf files is taking time

main_df.to_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/main_df_tokens.csv',index=False)
