==> notebooks/2.skills_from_json1.ipynb <==
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

#path for csv file generated in previous notebook
path='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/main_df_tokens.csv'
df=pd.read_csv(path)

df.head()

consider few cols in above df

new_df=df.loc[:][['id','category','tokens']]
new_df.head()



*   loading spacy model
*   adding entity ruler which has some json labeled data to recognise skills



import spacy
nlp = spacy.load("en_core_web_lg")
skill_pattern_path = "jz_skill_patterns.jsonl"

ruler = nlp.add_pipe("entity_ruler")
ruler.from_disk(skill_pattern_path)
nlp.pipe_names

def get_skills_tokens(tokens):
    tokens=tokens.split(',')
    text=' '.join(tokens)
    doc = nlp(text)
    skill_set=set()
    for ent in doc.ents:
        if ent.label_ == "SKILL":
            skill_set.add(ent.text)
    return ','.join(list(skill_set))


**these may take some time**

new_df['skills']=new_df['tokens'].apply(get_skills_tokens)

new_df.head()

getting skill count to know how many skills we got

new_df['skills_count_json1']=new_df['skills'].apply(lambda x:len(x.split(',')))

new_df.head()

save the csv file to avoid long computation

new_df.to_csv('skills_with_json1.csv',index=False)

new_df['skills_count_json1'][1586]

new_df['skills'][0]

set(df.category)
