==> notebooks/3.gen_new_skills.ipynb <==
import pandas as pd
import unicodedata

from google.colab import drive
drive.mount('/content/drive')

#path of csv where we got skills from json1
path1='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/skills_with_json1.csv'
df=pd.read_csv(path1)


df.head()

#path of main csv
path2='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/main_df.csv'
main_df=pd.read_csv(path2)
main_df.head()



*   aim is to train a unsupervised spacy model with complete resume text
*   from skills we got from json1. we generate new skills using model
*   we can get similar words from model
*   add it as s skills



import re

def preprocess(text):
    text = re.sub(r'[^\w\s\']',' ', text)
    text = re.sub(r'[ \n]+', ' ', text)

    #to remove circumflex(dataâ)
    text = unidecode.unidecode(text)
    return text.strip().lower()

main_df['pre_text']=main_df['text'].apply(preprocess)

main_df['text'][0][:900]

main_df['pre_text'][0][:900]

save as a text file,needed for training the unsupervised model

path3='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/models/complete.txt'
main_df.to_csv(path3, columns=["pre_text"], header=None, index=False)

import fasttext

model = fasttext.train_unsupervised(path3, "cbow")
#save the model
path4='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/models/fasttext_complete.bin'
model.save_model(path4)

model.get_nearest_neighbors('skills',k=16,)

model.get_nearest_neighbors('chef',k=16,)

getting all skills we got from json1-spacy model

all_skills=set()

for i in range(len(df)):
    if df.loc[i]['category']=='ACCOUNTANT':

        s=set(df.loc[i]['skills'].split(','))

        all_skills=s.union(all_skills)


len(all_skills)

i=87
model.get_nearest_neighbors(list(all_skills)[i],k=10),list(all_skills)[i]

reading json files , getting skills from it

import json

def read_json(path):
    output_data=[]
    with open(path, "r") as file:
        for line in file:
            data = json.loads(line)
            output_data.append(data)
    return output_data

def json_to_skills(path):

    json_list=read_json(path)
#     print(json_list)

    n=len(json_list)
    all_skills=set()
    s=[]
    for x in range(n):
        for i in range(len(json_list[x]['pattern'])):
            try:
                s.append(json_list[x]['pattern'][i]['LOWER'])
            except:
                s.append(json_list[x]['pattern'][i]['TEXT'].lower())


        string=' '.join(s)
        s=[]
        all_skills.add(string)
    return all_skills

main_json_skills=json_to_skills('jz_skill_patterns.jsonl')

list(main_json_skills)[963]

model.get_nearest_neighbors('never',k=16,)



*   json format to add skill
*   {"label":"SKILL","pattern":[{"LOWER":"agile"},{"LOWER":"project"},{"LOWER":"management"}]}





*   generate new skills from spacy trained model




def generate_new_skills(probability_threshold=0.87):
    cat=list(set(df['category']))
    new_gen_skills=[]
    for category in cat:
        all_skills=set()
        for i in range(len(df)):
            if df.loc[i]['category']=='ACCOUNTANT':
                s=set(df.loc[i]['skills'].split(','))
                all_skills=s.union(all_skills)


        for skill in list(all_skills):
            pred=model.get_nearest_neighbors(skill,k=10)
            new_set=set()
            for new in pred:
                dic={"label":"SKILL","pattern":[]}

                if new[0]<probability_threshold or new[1] in main_json_skills or new[1] in new_set:
                    continue
                new_set.add(new[1])

                for word in new[1].split():

                    dic['pattern'].append({"LOWER":word})

                new_gen_skills.append(dic)

    return new_gen_skills






new_gen_skills=(generate_new_skills())
# 0.7155979871749878<0.8

len(new_gen_skills)

writing a json files with newly generated skills

def write_json_with_list(l,path):
    with open(path, "w") as json_file:
    # Write the list of values to the JSON file
     for i in l:
        json.dump(i, json_file)
        json_file.write('\n')


#path to new jsonl file to add newly gen skills
path='/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/JsonFiles/new_gen_skill.jsonl'

write_json_with_list(new_gen_skills,path)

combining two json files

def merge_two_jsons(path1,path2,new_path):
    l1=read_json(path1)
    l2=read_json(path2)
    l=l1+l2

    write_json_with_list(l,new_path)




merge_two_jsons('jz_skill_patterns.jsonl','new_gen_skill.jsonl','total_skills.jsonl')
