==> notebooks/4.skills_from_json2.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd

main_df=pd.read_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/main_df_tokens.csv')

main_df.head()

loading the total skills json

!python -m spacy download en_core_web_lg

import spacy
nlp = spacy.load("en_core_web_lg")
skill_pattern_path = "/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/JsonFiles/total_skills2.jsonl"

ruler = nlp.add_pipe("entity_ruler")
ruler.from_disk(skill_pattern_path)
nlp.pipe_names

#save spacy model
nlp.to_disk('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/models/skills_extractor')

def get_skills_tokens(tokens):
    tokens=tokens.split(',')
    text=' '.join(tokens)
    doc = nlp(text)
    skill_set=set()
    for ent in doc.ents:
        if ent.label_ == "SKILL":
            skill_set.add(ent.text)
    return ','.join(list(skill_set))

skills_final=main_df.loc[:][['id','category','tokens']]
skills_final.head()

getting skills from tokens with final json
* it may take time

skills_final['skills']=skills_final['tokens'].apply(get_skills_tokens)

skills_final['skills_count_final']=skills_final['skills'].apply(lambda x:len(x.split(',')))

save the csv file

skills_final.to_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/skills_final.csv',index=False)

skills_final.head()

*  comparision between skills we got with different json files

skill_with_json1=pd.read_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/skills_with_json1.csv')

compare=pd.concat([skill_with_json1[['id','category','skills_count_json1']],skills_final[['skills_count_final']]],axis=1)

compare.head()

def count_range(x):
    if x==0:
        return 'zero'
    elif x>0 and x<=10:
        return 'very less 1-10'
    elif x>10 and x<=15:
        return 'less 11-15'
    elif x>15 and x<=25:
        return 'ok 16-25'
    elif x>25 and x<=35:
        return 'enough 26-35'
    else:
        return 'more >35'

compare['range_json1']=compare['skills_count_json1'].apply(count_range)
compare['range_json2']=compare['skills_count_final'].apply(count_range)


compare.head()

compare['range_json1'].value_counts()

compare['range_json2'].value_counts()


compare.head()

compare.to_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/compare.csv',index=False)
