==> notebooks/5.huggingface-jobdescription.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np

!pip install transformers
!pip install datasets

import transformers
import datasets

from datasets import load_dataset

Importing data

job_dataset = load_dataset('jacob-hugging-face/job-descriptions')

job_dataset

length=853

set(job_dataset['train']['position_title'])

taking four job description from software related

job_roles=['Software Developer','Software Engineer']

taking only
1.job_description 2.position_title

job_des=[]
pos_title=[]
indices=[]

#getting 2 job description of any position_tile

def get_job_des_pos(cat,job_des,pos_title):
    count=0

    for i in range(length):
        if job_dataset['train']['position_title'][i]==cat:
            job_des.append(job_dataset['train']['job_description'][i])
            pos_title.append(job_dataset['train']['position_title'][i])
            count+=1
            indices.append(i)
            if count==2:
                break



get_job_des_pos('Software Developer',job_des,pos_title)

len(job_des)

pos_title

get_job_des_pos('Software Engineer',job_des,pos_title)


pos_title

indices

selecting 15 job descriptions index numbers

l=list(np.random.randint(0,length, size=(15, )))

l

count=len(job_des)
for i in l:
    #to avoid already taken job descriptions
    if i not in indices:
            job_des.append(job_dataset['train']['job_description'][i])
            pos_title.append(job_dataset['train']['position_title'][i])

            count+=1
            if count==15:
                break


len(job_des),len(pos_title)

job_df=pd.DataFrame({'job_description':job_des,'position_title':pos_title})

job_df.head()

preprocess the job description and generate tokens

import re
import spacy
nlp=spacy.load('en_core_web_sm')

def preprocess(text):
    # convert all characters in the string to lower case
    text = re.sub(r'\n', ' ', text)

    #remove digits
    text = re.sub(r'\d', '', text)

    # Remove punctuation (you can customize this pattern for your needs)
    text = re.sub(r'[^\w\s]', '', text)

    # Remove extra spaces and replace with a single space
    text = re.sub(r'\s+', ' ', text).strip()

    # tokenize word
    doc=nlp(text)
    toks=[i.text.lower() for i in doc if not i.is_stop]


    return "," .join(toks)

job_df['tokens']=job_df['job_description'].apply(preprocess)

job_df.head()

getting skills from the tokens for each row

!python -m spacy download en_core_web_lg


import spacy
nlp = spacy.load("en_core_web_lg")
skill_pattern_path = "/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/JsonFiles/total_skills2.jsonl"

adding new entity(to spacy pipeline)

ruler = nlp.add_pipe("entity_ruler")
ruler.from_disk(skill_pattern_path)
nlp.pipe_names

def get_skills_tokens(tokens):
    tokens=tokens.split(',')
    text=' '.join(tokens)
    doc = nlp(text)
    skill_set=set()
    for ent in doc.ents:
        if ent.label_ == "SKILL":
            skill_set.add(ent.text)
    return ','.join(list(skill_set))

job_df['skills']=job_df['tokens'].apply(get_skills_tokens)

job_df['skills_count']=job_df['tokens'].apply(lambda x:len(x.split(",")))

job_df.head()

job_df['skills_count']

Save the csv file

job_df.to_csv('/content/drive/MyDrive/Colab Notebooks/resume skills - assessment/CsvFiles/job_df.csv',index=False)
